Remove package dependency from armbian-$RELEASE-desktop Closes https://github.com/armbian/build/issues/1544 Adjustement to armbian-config is needed.

+++ lib/desktop.sh
@@ -13,7 +13,7 @@ create_desktop_package ()
 {
 	# cleanup package list
 	PACKAGE_LIST_DESKTOP=${PACKAGE_LIST_DESKTOP// /,}; PACKAGE_LIST_DESKTOP=${PACKAGE_LIST_DESKTOP//[[:space:]]/}
-	PACKAGE_LIST_DESKTOP_RECOMMENDS=${PACKAGE_LIST_DESKTOP_RECOMMENDS// /,}; PACKAGE_LIST_DESKTOP_RECOMMENDS=${PACKAGE_LIST_DESKTOP_RECOMMENDS//[[:space:]]/}
+	PACKAGE_LIST_DESKTOP_SUGGESTS=${PACKAGE_LIST_DESKTOP_SUGGESTS// /,}; PACKAGE_LIST_DESKTOP_SUGGESTS=${PACKAGE_LIST_DESKTOP_SUGGESTS//[[:space:]]/}
 
 	local destination=${SRC}/.tmp/${RELEASE}/${BOARD}/${CHOSEN_DESKTOP}_${REVISION}_all
 	rm -rf "${destination}"
@@ -28,8 +28,8 @@ create_desktop_package ()
 	Installed-Size: 1
 	Section: xorg
 	Priority: optional
-	Depends: ${PACKAGE_LIST_DESKTOP//[:space:]+/,}
-	Recommends: ${PACKAGE_LIST_DESKTOP_RECOMMENDS//[:space:]+/,}
+	Recommends: ${PACKAGE_LIST_DESKTOP//[:space:]+/,}
+	Suggests: ${PACKAGE_LIST_DESKTOP_SUGGESTS//[:space:]+/,}
 	Provides: ${CHOSEN_DESKTOP}
 	Description: Armbian desktop for ${DISTRIBUTION} ${RELEASE}
 	EOF
@@ -52,9 +52,11 @@ create_desktop_package ()
 		fi
 
 		# Adjust menu
+		if [ -f /etc/xdg/menus/xfce-applications.menu ]; then
 		sed -i -n '/<Menuname>Settings<\/Menuname>/{p;:a;N;/<Filename>xfce4-session-logout.desktop<\/Filename>/!ba;s/.*\n/\
 		\t<Separator\/>\n\t<Merge type="all"\/>\n        <Separator\/>\n        <Filename>armbian-donate.desktop<\/Filename>\
 		\n        <Filename>armbian-support.desktop<\/Filename>\n/};p' /etc/xdg/menus/xfce-applications.menu
+		fi
 
 		# Hide few items
 		if [ -f /usr/share/applications/display-im6.q16.desktop ]; then mv /usr/share/applications/display-im6.q16.desktop /usr/share/applications/display-im6.q16.desktop.hidden; fi
